Document linting setup; rename kgroup1 -> kgroup
- Update CLAUDE.md, CHANGELOG.md, RELEASE_NOTES.md, README.md, and
  llms.txt to reflect the detekt + kotlinter linting setup,
  .editorconfig, Gradle configuration cache, readingbat 3.1.5 ->
  3.1.8, new Makefile targets, and shadowJar DuplicatesStrategy.WARN
- Rename the Kotlin challenge package kgroup1 -> kgroup and update
  Content.kt to match
- Surface stderr in test logging (STANDARD_ERROR event)
- Minor Content.kt formatting tidy-ups
