+++ CHANGELOG.md
@@ -7,16 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Fixed
+### Added
 
-- `make uberjar` (and the documented `./gradlew uberjar` command) was broken — no such task exists. Pointed `make uberjar` at `./gradlew buildFatJar` and updated `CLAUDE.md` / `llms.txt` accordingly.
+- Detekt (`dev.detekt` 2.0.0-alpha.3) and Kotlinter (`org.jmailen.kotlinter` 5.4.2) Gradle plugins for static analysis and ktlint formatting.
+- `.editorconfig` with project-wide formatting rules (UTF-8, LF, 2-space indent, 120-char max) and a curated set of disabled ktlint rules tuned to this codebase.
+- Makefile targets `lint`, `format`, `detekt`, `detekt-baseline`, and a self-documenting `help` target driven by `##` annotations on each target.
+- `_require-gradle-version` guard target that fails fast if `gradle/libs.versions.toml` parsing yields an empty version.
 
 ### Changed
 
-- Switched fat-jar configuration to Ktor's idiomatic `ktor { fatJar { archiveFileName = "server.jar" } }` block. `tasks.shadowJar` now only carries signature/license excludes (`META-INF/*.SF`, `*.DSA`, `*.RSA`, `LICENSE*`), collapsed into a single varargs `exclude(...)` call. This reverses the 1.7.0 consolidation toward `tasks.shadowJar` and removes parallel configuration paths.
+- Enabled Gradle configuration cache (`org.gradle.configuration-cache=true`).
+- Bumped `readingbat` 3.1.5 → 3.1.8.
+- Switched `tasks.shadowJar` to `DuplicatesStrategy.WARN` and dropped the `LICENSE*` exclude so duplicate-resource collisions are visible during the fat-jar build.
+- Hardened `Makefile` `GRADLE_VERSION` parsing with `grep -E` + `sed -E` instead of a single positional `sed -n` regex.
+- Narrowed the `io.ktor.server.testing.*` wildcard import in `ContentTests` to the specific `testApplication` symbol, per ktlint.
+- `make uberjar` (and the documented `./gradlew uberjar` command) was broken — no such task exists. Pointed `make uberjar` at `./gradlew buildFatJar` and updated `CLAUDE.md` / `llms.txt` accordingly.
+- Switched fat-jar configuration to Ktor's idiomatic `ktor { fatJar { archiveFileName = "server.jar" } }` block. `tasks.shadowJar` now only carries signature excludes (`META-INF/*.SF`, `*.DSA`, `*.RSA`). This reverses the 1.7.0 consolidation toward `tasks.shadowJar` and removes parallel configuration paths.
 - Promoted `jvm` (17) and `gradle` (9.5.0) versions into `gradle/libs.versions.toml`. `build.gradle.kts` reads the JVM toolchain via `libs.versions.jvm.get().toInt()`; the `Makefile`'s `upgrade-wrapper` target reads the Gradle version from the catalog via `sed`.
 - Renamed the version-catalog plugin alias `versions` → `ben-manes-versions` (the prior name collided with the `[versions]` table). Call site is now `alias(libs.plugins.ben.manes.versions)`.
 - Extracted repeated `"clean"` / `"build"` task-name strings to `cleanTask` / `buildTask` references in `build.gradle.kts`.
+
+### Fixed
+
+- Source files (`Content.kt`, `kgroup/IntLambda1.kt`, `kgroup/StringLambda1.kt`) now end with a trailing newline, per `.editorconfig`.
 
 ## [1.7.0] - 2026-05-03
