+++ RELEASE_NOTES.md
@@ -6,8 +6,14 @@ Build-housekeeping follow-ups to 1.7.0. No DSL or runtime behavior changes.
 
 ### Highlights
 
+- **Linting wired in.** Detekt (`dev.detekt` 2.0.0-alpha.3) and Kotlinter (`org.jmailen.kotlinter` 5.4.2) Gradle plugins are now applied, with matching `make lint` / `make format` / `make detekt` / `make detekt-baseline` shortcuts.
+- **`.editorconfig` landed.** Project-wide formatting rules (UTF-8, LF, 2-space indent, 120-char max, final newline) plus a curated set of disabled ktlint rules tuned to this codebase.
+- **Gradle configuration cache enabled.** `org.gradle.configuration-cache=true` in `gradle.properties`. New build logic should be configuration-cache-compatible.
+- **readingbat 3.1.5 → 3.1.8.** Picks up upstream fixes; no DSL changes required.
+- **Self-documenting Makefile.** `##` annotations on each target are surfaced via `make help`; a `_require-gradle-version` guard prevents silent wrapper-upgrade failures.
+- **ShadowJar duplicates surface.** `tasks.shadowJar` now uses `DuplicatesStrategy.WARN` (dropping the `LICENSE*` exclude) so duplicate-resource collisions in the fat jar are visible.
 - **Fixed broken fat-jar make target.** `make uberjar` (and the previously documented `./gradlew uberjar`) referenced a task that does not exist. The Makefile now invokes `./gradlew buildFatJar`; `CLAUDE.md` and `llms.txt` updated to match.
-- **Single fat-jar configuration path.** Switched to Ktor's idiomatic `ktor { fatJar { archiveFileName = "server.jar" } }` block; `tasks.shadowJar` now only carries the META-INF signature/license excludes. Reverses the 1.7.0 direction (consolidation onto `tasks.shadowJar`) in favor of the Ktor-recommended API.
+- **Single fat-jar configuration path.** Switched to Ktor's idiomatic `ktor { fatJar { archiveFileName = "server.jar" } }` block; `tasks.shadowJar` now only carries the META-INF signature excludes. Reverses the 1.7.0 direction (consolidation onto `tasks.shadowJar`) in favor of the Ktor-recommended API.
 - **JVM and Gradle versions now in the catalog.** `gradle = "9.5.0"` and `jvm = "17"` are tracked under `[versions]` in `gradle/libs.versions.toml`. `build.gradle.kts` reads `libs.versions.jvm.get().toInt()` for the toolchain; `Makefile` reads `gradle` from the catalog when running `upgrade-wrapper`.
 - **Catalog plugin alias renamed.** `versions` → `ben-manes-versions` to avoid the name collision with the `[versions]` table. Call site is `alias(libs.plugins.ben.manes.versions)`.
 
@@ -15,12 +21,15 @@ Build-housekeeping follow-ups to 1.7.0. No DSL or runtime behavior changes.
 
 - Forks invoking `./gradlew uberjar` should switch to `./gradlew buildFatJar` (or use `make uberjar`).
 - Forks referencing `libs.plugins.versions` should rename to `libs.plugins.ben.manes.versions`.
+- Run `make lint` (or `./gradlew lintKotlin detekt`) before submitting changes — CI/local builds will surface style and static-analysis violations.
 
 ### Verification
 
-- `./gradlew buildFatJar` — produces `build/libs/server.jar` (~184 MB).
+- `./gradlew buildFatJar` — produces `build/libs/server.jar`.
+- `./gradlew lintKotlin detekt` — both checks run cleanly on `master`.
 - `./gradlew tasks` and `./gradlew help` resolve cleanly.
 - `make -n upgrade-wrapper` expands to `./gradlew wrapper --gradle-version=9.5.0 --distribution-type=bin`.
+- `make help` lists every documented Makefile target.
 
 ---
