+++ src/main/kotlin/kgroup/StringLambda1.kt
@@ -1,4 +1,4 @@
-package kgroup1
+package kgroup
 
 // @desc Description of **combine()**
